requirements.txt
asgiref==3.3.1
astroid==2.4.2
Django==3.1.4
django-countries==7.0
django-crispy-forms==1.10.0
django-phone-field==1.8.1
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.0.1
psycopg2-binary==2.8.6
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
python-decouple==3.3
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
whitenoise==5.2.0
wrapt==1.12.1
